requirements.txt
asgiref==3.2.10
Babel==2.8.0
Django==3.0.9
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-phonenumber-field==4.0.0
gunicorn==20.0.4
phonenumberslite==8.12.7
Pillow==7.2.0
psycopg2-binary==2.8.5
pytz==2020.1
PyYAML==5.3.1
sqlparse==0.3.1
